requirements.txt
allure-pytest==2.13.5
allure-python-commons==2.13.5
attrs==25.3.0
certifi==2025.1.31
charset-normalizer==3.4.1
greenlet==3.1.1
idna==3.10
iniconfig==2.1.0
packaging==24.2
playwright==1.51.0
pluggy==1.5.0
pyee==12.1.1
pytest==8.3.5
python-dotenv==1.1.0
requests==2.32.3
typing_extensions==4.13.0
urllib3==2.3.0
